[Silver III] Title: 주유소, Time: 360 ms, Memory: 36068 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,8 +8,8 @@ public class Main {
         BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(System.in));
 
         int n = Integer.parseInt(bufferedReader.readLine());
-        int[] gasStation = new int[n];
-        int[] distance = new int[n-1];
+        long[] gasStation = new long[n];
+        long[] distance = new long[n-1];
 
         StringTokenizer stringTokenizer = new StringTokenizer(bufferedReader.readLine());
         for (int i = 0; i < n-1; i++) {
@@ -21,8 +21,8 @@ public class Main {
             gasStation[i] = Integer.parseInt(stringTokenizer.nextToken());
         }
 
-        int cost = 0;
-        int current = gasStation[0];
+        long cost = 0;
+        long current = gasStation[0];
 
         for (int i = 0; i < n-1; i++) {
             if (gasStation[i] < current) {
